backend/requirements.txt
# FastAPI and server
fastapi==0.100.0
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
pydantic==2.11.4
pydantic-settings==2.6.1

# Azure OpenAI for APIM integration
openai==2.26.0
azure-identity==1.15.0

# HTTP client
httpx==0.24.1

# CORS and middleware
python-multipart==0.0.9

# Logging
python-json-logger==2.0.7
